subtask(F-5-2-split-install-manifest-helper-modules): Split install manifest helper file into focused modules
