+++ crates/tanren-bdd/src/steps/install/manifest_helpers/hash.rs
@@ -1,0 +1,5 @@
+use tanren_cli_app::install::manifest::sha256_hex;
+
+pub(crate) fn sha256_hex_string(bytes: &[u8]) -> String {
+    sha256_hex(bytes)
+}
